--- oldsrc/representor/index.ts
@@ -1,17 +1,0 @@
-import Representor from './base';
-import CollectionJsonRepresentor from './collection-json';
-import HalRepresentor from './hal';
-import RepresentorHelper from './helper';
-import HtmlRepresentor from './html';
-import JsonApiRepresentor from './jsonapi';
-import SirenRepresentor from './siren';
-
-export {
-  CollectionJsonRepresentor,
-  HalRepresentor,
-  HtmlRepresentor,
-  JsonApiRepresentor,
-  Representor,
-  RepresentorHelper,
-  SirenRepresentor
-};
