+++ src/client.ts
@@ -6,7 +6,8 @@ import { factory as binaryState } from './state/binary';
 import { factory as jsonApiState } from './state/jsonapi';
 import { factory as sirenState } from './state/siren';
 import { factory as textState }from './state/text';
-import { factory as cjState }from './state/collection-json';
+import { factory as cjState } from './state/collection-json';
+import { factory as htmlState } from './state/collection-json';
 import { parseContentType } from './http/util';
 import { resolve } from './util/url';
 import { LinkVariables } from './link';
@@ -25,6 +26,7 @@ export default class Client {
     'application/vnd.siren+json': [jsonApiState, '0.9'],
     'application/vnd.collection+json': [cjState, '0.9'],
     'application/json': [halState, '0.8'],
+    'text/html': [htmlState, '0.7'],
   }
 
   constructor(bookmarkUri: string) {
