+++ src/types/uri-template.d.ts
